requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-q>=1.3,<2.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<22.0